requirements.txt
#utilities
tqdm
itertools

#data
pickle
pandas

#plotting
matplotlib
seaborn

#scientific computing
numpy
scipy

#machine learning
torch
torchsummary

#graphs
networkx
torch_geometric
